# src/main/java/com/example/quote/api/dto/FieldErrorDto.java
package com.example.quote.api.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class FieldErrorDto {

    private String field;
    private String message;
    private Object rejectedValue;
    private String code;
}