# src/main/java/com/example/quote/api/dto/QuoteRequest.java
package com.example.quote.api.dto;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.time.OffsetDateTime;
import java.util.Map;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class QuoteRequest {

    @NotBlank
    @Size(max = 64)
    private String documentId;

    @NotBlank
    @Size(max = 64)
    private String customerId;

    @NotBlank
    @Size(min = 3, max = 3)
    private String currency;

    @NotNull
    @DecimalMin(value = "0.0", inclusive = true)
    private BigDecimal amount;

    @Min(0)
    @Max(1)
    private Double riskScore;

    @NotNull
    private OffsetDateTime effectiveFrom;

    private OffsetDateTime effectiveTo;

    private Map<String, String> metadata;
}